split tasks and configuration holders into new classes
